requirements.txt
flask 
flask-sqlalchemy 
flask-login 
flask-wtf 
mysql-connector-python 
python-dotenv
